Add new KMS key policy statements for admin and AWS Config
This commit includes a new policy statement "AllowAdminToRoot" to grant administrative actions to the account root user. Additionally, it adds a "sid" for AWS Config in the existing statement and uses the `aws_caller_identity` data source to dynamically fetch the account ID. These changes improve flexibility and security configuration.